Update README.md
Added some instructions on creating a bash script that might help other Mac users, but I'm sure it can also be adapted easily for Linux. We've not discussed how to manage things like this so I've created it as a new branch and will make a PR so you can decide whether to use it as-is or whether you want to include this in a different way.