statsWrapper can now process and return uncertainties, but how it does it may not be ideal.
For mean, we disregard incoming SDs and output the SD of the nominal values.

For sum, we output the SD of the sum, which should be the root of the sum of the squares of the individual SDs. This means that the SD of the sum of a lot of different values with zero SD will be zero.

For other stats functions (min, max) we disregard incoming SDs and output zero SD.

Minor change to function help indicating that "mean" and "sd" disregard uncertainty values in their inputs.